Renames Sandbox namespace to Map.  Also cleans up project file.
