+++ ParquetClassLibrary/Map/MapRegionEditor.cs
@@ -1,11 +1,11 @@
 using System;
 using System.Collections.Generic;
 using System.IO;
-using ParquetClassLibrary.Sandbox.Parquets;
+using ParquetClassLibrary.Parquets;
 using ParquetClassLibrary.Stubs;
 using ParquetClassLibrary.Utilities;
 
-namespace ParquetClassLibrary.Sandbox
+namespace ParquetClassLibrary.Map
 {
     // TODO: Undo/Redo System
     // TODO: ParquetSelect, ParquetCopy, ParquetPaste, ParquetClear functionality
